--- nubilanding/src/componentes/beneficios/styles.module.scss
@@ -1,50 +1,0 @@
-@import url("https://fonts.googleapis.com/css?family=Roboto:400,500|Rubik:300,400,500,500i&display=swap");
- 
-.contenedor {
-  /*no tiene que tener margin*/
-  width: 100%;
-  height: 50%;
-  border: 50px solid transparent;
-  background-color: #EEEEEF;
-  
-  
-}
- 
-.Contenedor {
-    
-  width: 25%; /*el ancho esta bien en web*/
-  height: 250%;
-  border: green solid 2px;
-  text-align: left;
-  
- 
-}
- 
-.titulo{
-  color:#1D4A62;
-  font-family: 'Rubik';
-  font-size: 36px;
-}
-.guion{
-  color:#00FCCB;
-  font-size: 18px;
-  
-}
- 
-.texto{
-  color:#5E737E;
-  font-size: 18px;
-  font-family: 'Roboto';
-}
-
-/*.boton{
-  background-color: #6159FF;
-  color:#FAF3FD;
-}
-subtitulo:
-guion: 
-parrafo:
-fondo:
-boton-fondo:
-letra-boton:
-*/
