+++ nubilanding/src/components/beneficios/styles.module.scss
@@ -1,0 +1,64 @@
+@import url("https://fonts.googleapis.com/css?family=Roboto:400,500|Rubik:300,400,500,500i&display=swap");
+ 
+
+
+.contenedorGeneral{
+  width: 100%;
+  height: 200%;
+  background-color: #EEEEEF;
+
+}
+ 
+/*al .Contenedor genral ponerle weight de 100% para que los 4 c tengan el mismo espacio ya esta*/
+.contenedor {
+  /*no tiene que tener margin*/
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-wrap:wrap;
+  justify-content: center;
+  
+  }  
+
+.Contenedor {
+    
+  width: 30%; /*el ancho esta bien en web*/
+  height: 250%;
+  text-align: justify; 
+  margin: 5px 50px 30px 25px;
+  padding: 15px;
+  
+ 
+}
+
+ 
+.titulo{
+  color:#1D4A62;
+  font-family: 'Roboto'; 
+  font-weight: 500; /*esto es para especificar ? grosor*/
+  font-size: 36px; 
+  font-style: italic;
+}
+.guion{
+  color:#00FCCB;
+  font-size: 36px;
+  margin:  -50px auto -30px; /*solo arriba y abajo*/ 
+  
+} 
+ 
+.texto{
+  color:#5E737E;
+  font-size: 18px;
+  font-family: 'Roboto'; 
+}
+
+.boton{
+  font-size:14px;
+  background-color: #6159FF;
+  color:#FAF3FD;
+  border: transparent solid 1px;
+  border-radius: 30px;
+  padding: 20px 30px;
+  
+ 
+}
